// src/lib/enum.tsx
export enum Categories {
  Ambitious = "Ambitious",
  HeartAndHome = "Heart & Home",
  SelfLove = "Self Love",
  SoulfulImpact = "Soulful Impact",
}

export enum Category {
  Motivation = "Motivation",
  Obstacle = "Obstacle",
}

export enum Motivation {
  CareerWork = "Career/Work",
  Family = "Family",
  RelationshipFriends = "Relationship/Friends",
  SharingContributions = "Sharing/Contributions",
  SelfDevelopment = "Self-Development",
  FinanceMoney = "Finance/Money",
  Health = "Health",
  Spiritual = "Spiritual",
}

export enum Season {
  InspiringFlames = "Inspiring Flames",
  LifesCanvas = "Life’s Canvas",
  HeartAndHome = "Heart & Home",
  RibbonsAndPaperBoxes = "Ribbons and Paper Boxes",
  TrophiesAndTriumphs = "Trophies and Triumphs",
  ARefilledHeart = "A Refilled Heart",
}

export enum Chapter {
  Start = "start",
  Middle = "middle",
  End = "end",
}

export enum Obstacle {
  NotGoodEnough = "NotGoodEnough",
  NotPrioritize = "NotPrioritize",
  NotLoveMyself = "NotLoveMyself",
  FearNotAccepted = "FearNotAccepted",
  TimeManagement = "TimeManagement",
  LackConfidence = "LackConfidence",
  LackInspiration = "LackInspiration",
  OverDriven = "OverDriven",
  SocietyPressure = "SocietyPressure",
  OverwhelmedWorld = "OverwhelmedWorld",
}